changed the size field in RCPackageSpec to installed_size. Added a boolean
* src/common.c, src/common.h: changed the size field in RCPackageSpec
to installed_size.  Added a boolean installed flag to the spec, and
support for this to the init function.  Fixed the prototype for
rc_package_spec_free to rc_package_spec_free_members.

* src/rc-packman.c, src/rc-packman.h, src/rc-packman-private.h: first
commit for the new packman object, which uses the "new" API we worked
out.  Currently no working backends, so... ;)  The only thing that
should ever use rc-packman-private.h is the backends, so in other words
don't use it.