in all of the virtual functions, zero out the error fields
* src/rc-packman.c: in all of the virtual functions, zero out the error
fields (RC_PACKMAN_NONE and NULL), so that we'll know when they return
where an error actually occured.  D'oh.

* src/common.c (rc_mktmpdir): fix a slight oversight where we assumed
the prefix ended in '/', leading to directory names like
/tmpredcarpet-XXXXXX instead of /tmp/redcarpet-XXXXXX.

* src/rc-packman.c: killed all of the rc_packman_fake_* functions.
From now on, if you try to call a function which isn't implemented by a
backend, you'll trigger an assertion, but since this is the kind of
thing which should /never/ happen, it's a good idea.  Edited the init
and class init functions appropriately.