implemented a simple locking mechanism around some of the functions which
* src/rc-packman.h, src/rc-packman.c: implemented a simple locking
mechanism around some of the functions which may return control to the
gtk main loop in the middle, as a safety precaution because I'm
paranoid.  If the RC_PACKMAN_BUSY error is ever triggered we should
probably look very carefully at what caused it.  Also got rid of the
unused status enum.