reworked slightly the way errors are handled, to make it more consistant
* src/rc-packman.h, src/rc-packman.c, src/rc-debman.c, src/rc-rpmman.c:
reworked slightly the way errors are handled, to make it more
consistant and hopefully usable.  Now all errors are reported
to the object internally, and the error code should be checked after
every rc_packman call with rc_packman_get_error (and _get_reason, if
necessary).  That's it.

* src/softmgr-install.c: fixed the one! place in our code that checks
errors so far.  Errm...

* src/rc-rpmman.c: fixed joe's bug with the non-dup'ed strings.  I
think.  I really need to get a new rh testing machine.