added functions to do the installation in the gui with the pretty dialog
* src/gui-install.c, src/gui-install.h: added functions to do the
installation in the gui with the pretty dialog and all

* src/Makefile.am: added gui-install.[ch] to the gui sources

* src/gui.c: added gui-install.h to includes

* src/gui.h: include glade.h and gtkhtml.h, since we use them and can't
depend on only being included by gui.c

* src/rc-debman.c: every time we fork, we need to reset the SIGPIPE
signal to SIG_DFL.  Finding this out sucked, and I'm really pissed at
dpkg for not taking care of this itself.
(rc_debman_install): remember the total number of packages to install,
and pass this along when we emit the pkg_installed signal.

* src/rc-packman.h, src/rc-packman.c: changed the pkg_installed signal
to always report the total number of packages being installed, to make
it easier to use a progress bar.  What a shitty api.  Updated the
emission wrapper, marshaller, etc.

* src/rc-rpmman.h, src/rc-rpmman.c: added an internal object variable
to keep track of the number of packages installed and the number total
in the transaction, and changed all calls to
rc_packman_package_installed to match the changed signal.