rewrote this function. Instead of unpacking the packages one fork at a
* src/rc-debman.c (rc_debman_install): rewrote this function.  Instead
of unpacking the packages one fork at a time (which was slow as balls
because of a) forking and b) dpkg reading the database), we instead
unpack them in one call to dpkg.  To be able to send installed signals,
I open a pipe between the forks, have dpkg write to the pipe instead
of stdout, and parse the output of dpkg to determine when a package has
been installed.  It's pretty slick.