changed some of the category mappings
* src/rc-debman.c: changed some of the category mappings

* src/gui-install.c: fix install with deps, the code sucks and I am in
a bad mood and I hope it works perfectly because the code makes me
sick.  BTW refreshing channels doesn't work correctly because it's not
refreshing all of the channels, for some reason, and since we're now
installing and removing from multiple channels... if you want to check
it out, on the current testing server, try installing gnucash on
debian from the helix gnome channel, it should pull things in from
the debian channel.