Preliminary code to get nicely-displayable dep info (packages X, Y, Z will
* libredcarpet/deps-private.h, libredcarpet/deps.c, libredcarpet/deps.h:
Preliminary code to get nicely-displayable dep info (packages X, Y, Z will be
brought in to satisfy package A that was requested, packages D and E for pkg B, etc).
Will require more hacking to get this info out.

* libredcarpet/rc-package-dep.c: Made verify return FALSE if the dep was < or <=
and the target had no version info. This was to get around an issue where a package
conflicted with < V of another package bar, and bar was a virtual package provided
by bar1.2. This should be investigated further.

* libredcarpet/test_deps.c: Added a few features; this will be used now to verify
helix packages against RC before pushing to server.

* tools/distcheck.pl: Really added it this time. :-)

* tools/make_pkgs.c: Handling of &amp. seems to have gotten dropped out of previous
commit.