Changed the read_line function to use a line buffer for reading. Lots of
* libredcarpet/rc-debman.c: Changed the read_line function to
use a line buffer for reading. Lots of speed improvements. This
required changes to other code, as there was a second data
structure, RCLineBuf *, that had to be used.

* libredcarpet/rc-util.c, libredcarpet/rc-util.h: Added the
RCLineBuf stuff

* libredcarpet/rc-string.c, libredcarpet/rc-string.h: Added string
append functions

* libredcarpet/pkginfo.c: Went back to using strncmp's instead of
casecmps for comparing stuff. This might break if someone uploads
lowercase stuff to the debian packages file (as has happened)