Added lots of code to handle some cases that really, really, really need
* libredcarpet/deps.c: Added lots of code to handle some cases that
really, really, really need testing. If package A is installed and
it depends on B == 1.0, and you try to install package C which depends
on any version of A and upgrade	A and B in the same operation, this is
what used to happen: the A dep was processed from the system (because C
was seen first) and its hard dep on B was put in. When the new A was
processed, the hard dep on B would fail (notwithstanding the new B). There
are two solutions -- one was to remove all traces of the old package; this
is what untangle_package tries to do. Note that this is probably not a safe
thing to do, and now it's only called in cases where nothing else can be
done. The other thing, which is what's done, is to be a bit more intelligent
about it, by checking to see if something you're depending on is in the
installed list, and processing that (as long as it resolves other deps).
This needs testing; right now, the code seems to return a few more packages
than apt does.

* libredcarpet/rc-package.c (rc_package_compare): Made it just call
rc_package_spec_compare -- it wasn't checking for null versions/revisions.