added. This bad boy is a GtkObject that provides buffered read-a-line from
* libredcarpet/rc-line-buf.c, libredcarpet/rc-line-buf.h: added.
This bad boy is a GtkObject that provides buffered read-a-line from a
fd (aka a file or a pipe) using g_io_add_watch stuff.  I totally like
my little object, it's pretty dope.

* libredcarpet/Makefile.am: added rc-line-buf.[ch] to the sources for
libredcarpet

* libredcarpet/rc-debman.c: so some fixicating and an attempt to add
the "watch for stupid needy-of-manual-configuring packages" feature
turned into a total rewrite (almost) of this guy to make that doable.
So now literally everything in rc-debman uses rc-line-buf, so that it
can operate from within a GMainLoop I run by hand, which lets me use
g_io_add_watch to watch for a dpkg-run script trying to read... which
is what I've got to do next.  But it's now doable.
Also, add the feature that joe requested -- it now emits
transaction_step and configure_step with a seqno of 0 before it does
any transactioning or configuring, so that the gui can be smarter

* libredcarpet/rc-rpmman.c: add the same new feature as in debman (the
0 seqno signals)

* libredcarpet/rc-util.c, libredcarpet/rc-util.h: remove all traces of
the old RCLineBuf (sorry vlad!)

* libredcarpet/packman_test.c: fix an odd crash or two, add command
completion ;)