Change same_as into a pretend_name -- it's only to be used if we do not
* libredcarpet/rc-distro.c, libredcarpet/rc-distro.h: Change
same_as into a pretend_name -- it's only to be used if we do not
support the other distro uniquely, as it will break things if
there is a channel for both the pretend name and the real name. In
fact, this should probably not be used at all.

* libredcarpet/gui-init.c: Change some of the startup channel
parsing bits to reflect rc-distro changes.