Add rc_line_buf_get_buf which gets the current internal buffer even if
* libredcarpet/rc-line-buf.h, libredcarpet/rc-line-buf.c:
Add rc_line_buf_get_buf which gets the current internal buffer even if
there's been no '\n' yet.  Needed for The Debian Hack.

* libredcarpet/rc-debman.c: OH MY GOD IT WORKS!
I'm not going to try to give you a good changelog entry on this,
because I had to do so many crufty things it's not even funny.  But it
works.  This won't work until Vlad commits dpkg-helper.c and the
associated Makefile.am loving.