Changed the #define's to enums, to make for cleaner code and also so that
* libredcarpet/rc-package-dep.c, libreadcarpet/rc-package-dep.h:
Changed the #define's to enums, to make for cleaner code and also
so that gdb can give us names insted of numbers. Added
rc_packge_dep_verify_and_relation, which verifies a
RCPackageDep as if it were AND'd instead of OR'd. Also added
rc_package_dep_item_is_subset, which will attempt to merge
two RCPackageDepItems.