Do a bit of sanity checking, make sure we never add an anonymous package
* rc-deps.c (deps_add_anonymous_package): Do a bit of sanity
checking, make sure we never add an anonymous package that already
exists in some other form.
(deps_find_provide): don't assume that just because we found a
newer package for fixing a pdi that it will indeed fix it -- check
to make sure that package hasn't already been added to the list.

PLEASE TEST ME