another shared object, works with RPM 4.0.3
* src/rc-rpm-helper-with-rpmio-and-rpmdb.so: another shared
object, works with RPM 4.0.3

* src/Makefile.am: dist and install
rc-rpm-helper-with-rpmio-and-rpmdb.so

* src/rc-rpmman.c, src/rc-rpmman.h: conditionalize a lot of stuff
on both the API and ABI version using both the preprocessor and
runtime checks.  Add the third shared object to the list of things
we attempt to dlopen.

* configure.in: added --enable-rpm-version, which accepts a dotted
triple, and is used to make sure we compile something the compiler
will accept, regardless of what ABI we intend to run against.  For
cases like RPM 4.0, use 4.0.0 instead.